Exercicios 09 e 10

+++ Exercicios/ProgDesc/Ponteiros/EX08/ex08.c
@@ -6,7 +6,7 @@ Imprima o endereco de cada posicao desse array.
 #include <stdio.h>
 
 void main() {
-    int x[10];
+    float x[10];
 
     
     for(int i = 0; i < 10; i++){
